main/requirements.txt
numpy==1.23.5
numpy-quaternion==2022.4.2
